Working for 2 levels, but not for arbitrary levels

+++ graph_wrapper.py
@@ -6,19 +6,26 @@ import relaxmap
 class GraphWrapper(object):
 
     def __init__(self, vertices, edges, child=None):
+        self.parent = None
         self.child = child
+        #TODO: mk vertices optional
         edges = edges.join(vertices, {'__dst_id':'__id'})
-        edges.remove_column('community_id')
+        if 'community_id' in edges.column_names():
+            edges.remove_column('community_id')
         self.g = gl.SGraph(vertices=vertices, edges=edges, 
             vid_field='__id', src_field='__src_id', dst_field='__dst_id')
 
     AVG_COMM_SIZE = 25
 
     def homes_for_the_homeless(self):
-        comm = self.g.vertices['community_id']
-        if comm.num_missing() > (self.g.vertices.num_rows() * .5):
-            raise Exception('not implemented yet')
-        unique = list(comm.unique())
+        unique = [x for x in list(self.g.vertices['community_id'].unique()) if x]
+        if not unique:
+            unique = ['1']
+        if self.g.vertices['community_id'].num_missing() > (self.g.vertices.num_rows() * .5):
+            distance = (self.g.vertices.num_rows() / self.AVG_COMM_SIZE) - len(unique)
+            if distance > 0:
+                offset = int(max(unique)) + 1
+                unique += [str(x+offset) for x in range(int(distance))]
         self.g.vertices['community_id'] = self.g.vertices['community_id'].apply(lambda x: 
             random.choice(unique) if not x else x, skip_undefined=False)
 
@@ -26,15 +33,15 @@ class GraphWrapper(object):
         """"""
 
     def get_community_vertices(self):
-        comm = self.g.vertices['community_id']
-        unique = list(comm.unique())
-        if len(unique) < self.AVG_COMM_SIZE:
-            raise Exception('not implemented yet')
-        super_comms = [str(x) for x in range(len(unique) / self.AVG_COMM_SIZE)]
-        super_comms = [random.choice(super_comms) for x in unique]
-        return gl.SFrame({'__id': unique, 'community_id': super_comms})
+        unique = list(self.g.vertices['community_id'].unique())
+        num_comms = len(unique) / self.AVG_COMM_SIZE if len(unique) > self.AVG_COMM_SIZE else 1
+        num_comms = int(num_comms)
+        comms = [str(x) for x in range(num_comms)]
+        comms = [random.choice(comms) for x in unique]
+        return gl.SFrame({'__id': unique, 'community_id':comms})
 
     def get_community_edges(self):
+        #TODO: suspect this can be done more efficiently w/ triple_apply
         edges = self.g.edges.join(self.g.vertices, {'__dst_id':'__id'})
         edges.rename({'community_id':'dst_community_id'})
         edges = edges.join(self.g.vertices, {'__src_id':'__id'})
@@ -52,11 +59,11 @@ class GraphWrapper(object):
 
     def get_community_gw(self):
         self.homes_for_the_homeless()
-        return GraphWrapper(self.get_community_vertices(), 
-            self.get_community_edges(), child=self)
+        gw = GraphWrapper(self.get_community_vertices(), self.get_community_edges(), child=self)
+        self.parent = gw
+        return gw
 
     def find_communities(self, partitions=None):
-
         if not partitions:
             partitions = [self.g]
 
@@ -65,24 +72,25 @@ class GraphWrapper(object):
         mdl = 0
         vertices = None
         for partition in partitions:
-            print 'use partition v count: %s' % partition.vertices.num_rows()
+            print 'use partition vertex count: %s' % partition.vertices.num_rows()
             _vertices, _mdl = relaxmap.find_communities(partition)
             mdl += _mdl
             if not vertices:
                 vertices = _vertices
             else:
-                _offset = vertices.num_rows()
+                _offset = len(vertices['community_id'].unique())
                 _vertices['community_id'] = _vertices['community_id'].apply(
                     lambda x: str(int(x) + _offset))
                 vertices = vertices.append(_vertices)
-        
-        print 'vertices.num_rows %s' % vertices.num_rows()
-        print 'self.g.vertices.num_rows %s' %  self.g.vertices.num_rows()
+                assert len(vertices['community_id'].unique()) - len(_vertices['community_id'].unique()) == _offset
 
-        #update self.g.vertices w/ community info
-        vertices.sort('__id')
-        self.g.vertices.sort('__id')
-        self.g.vertices['community_id'] = vertices['community_id']
+        self.g = gl.SGraph(vertices=vertices, edges=self.g.get_edges(), 
+            vid_field='__id', src_field='__src_id', dst_field='__dst_id')
+        self.homes_for_the_homeless()
+
+        if self.parent:
+            self.parent.g = gl.SGraph(edges=self.get_community_edges(), 
+                src_field='__src_id', dst_field='__dst_id')
         
         #pull sgraph partitions out of child
         if self.child:
@@ -90,11 +98,17 @@ class GraphWrapper(object):
             grouped_vertices = vertices.groupby('community_id', {'__ids': gl.aggregate.CONCAT('__id')})
             for row in grouped_vertices:
                 vertices = gl.SFrame({'__id':row['__ids']})
-                print 'mk partition v count: %s' % vertices.num_rows()
+                print 'mk partition vertex count: %s' % vertices.num_rows()
                 p_vertices = self.child.g.vertices.join(vertices, {'community_id':'__id'})
+                if not p_vertices.num_rows():
+                    continue
                 p_edges = self.child.g.edges.join(p_vertices, {'__src_id':'__id'})
+                if not p_edges.num_rows():
+                    continue
                 p_edges.remove_column('community_id')
                 p_edges = p_edges.join(p_vertices, {'__dst_id':'__id'})
+                if not p_edges.num_rows():
+                    continue
                 p_edges.remove_column('community_id')
                 partitions.append(gl.SGraph(vertices=p_vertices, edges=p_edges, 
                     vid_field='__id', src_field='__src_id', dst_field='__dst_id'))
@@ -103,7 +117,7 @@ class GraphWrapper(object):
 
     def save(self):
         """
-        Call save on all childs as well
+        Call save on all children as well
         """
 
     @classmethod
